requirements.txt
semantic-kernel==1.3.0
pyautogen==0.2.25
httpx==0.27.0
pyyaml==6.0.1
python-dotenv==1.0.0
# Azure AI Search (optional RAG)
azure-search-documents==11.6.0b4
azure-core==1.30.1
# Anthropic (optional Claude provider)
anthropic==0.34.2
